feat(hub): update application bootstraping
* use `bootstrapApplication` from `@ovh-ux/manager-core`
* replace preloader by `@ovh-ux/manager-preloader`
