+++ packages/manager/apps/hub/.eslintrc.json
@@ -2,7 +2,6 @@
   "globals": {
     "__NODE_ENV__": true,
     "__VERSION__": true,
-    "__WEBPACK_REGION__": true,
     "moment": true,
     "__NG_APP_INJECTIONS__": true
   }
