+++ packages/manager/apps/hub/package.json
@@ -10,10 +10,7 @@
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "scripts": {
-    "build": "yarn build:eu && yarn build:ca && yarn build:us",
-    "build:ca": "webpack --env.production --env.region='CA' && mv dist dist-CA",
-    "build:eu": "webpack --env.production && mv dist dist-EU",
-    "build:us": "webpack --env.production --env.region='US' && mv dist dist-US",
+    "build": "webpack --env.production",
     "dev": "webpack-dev-server",
     "dev:watch": "yarn run dev",
     "start": "lerna exec --stream --scope='@ovh-ux/manager-hub-app' --include-dependencies -- npm run build --if-present",
@@ -30,6 +27,7 @@
     "@ovh-ux/manager-models": "^1.0.0",
     "@ovh-ux/manager-navbar": "^2.2.0 || ^3.0.0",
     "@ovh-ux/manager-ng-layout-helpers": "^0.0.0 || ^1.0.0",
+    "@ovh-ux/manager-preloader": "^0.0.0 || ^1.0.0",
     "@ovh-ux/ng-at-internet": "^5.0.0",
     "@ovh-ux/ng-at-internet-ui-router-plugin": "^3.0.0",
     "@ovh-ux/ng-ovh-api-wrappers": "^4.0.7",
@@ -62,7 +60,8 @@
     "oclazyload": "^1.1.0",
     "ovh-api-services": "^9.42.0",
     "ovh-ui-angular": "^3.15.1",
-    "ovh-ui-kit": "^2.40.0"
+    "ovh-ui-kit": "^2.40.0",
+    "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
     "@ovh-ux/manager-webpack-config": "^3.3.0",
