+++ packages/manager/apps/hub/src/app.module.js
@@ -1,0 +1,120 @@
+import 'script-loader!moment/min/moment-with-locales.min.js'; //eslint-disable-line
+
+import { Environment } from '@ovh-ux/manager-config';
+import angular from 'angular';
+import 'angular-translate';
+import uiRouter, { RejectType } from '@uirouter/angularjs';
+import ngOvhUiRouterLineProgress from '@ovh-ux/ng-ui-router-line-progress';
+import ngUiRouterBreadcrumb from '@ovh-ux/ng-ui-router-breadcrumb';
+
+import isString from 'lodash/isString';
+
+import 'ovh-ui-angular';
+import ovhManagerCore from '@ovh-ux/manager-core';
+import ovhManagerHub from '@ovh-ux/manager-hub';
+import ovhManagerNavbar from '@ovh-ux/manager-navbar';
+import ovhManagerOrderTracking from '@ovh-ux/ng-ovh-order-tracking';
+import { detach as detachPreloader } from '@ovh-ux/manager-preloader';
+
+import get from 'lodash/get';
+import has from 'lodash/has';
+import head from 'lodash/head';
+
+import atInternet from './components/at-internet';
+import errorPage from './components/error-page';
+import dashboard from './dashboard';
+
+import { BILLING_REDIRECTIONS } from './constants';
+
+import controller from './controller';
+import routing from './routing';
+import 'ovh-ui-kit/dist/oui.css';
+import './index.less';
+import './index.scss';
+
+Environment.setVersion(__VERSION__);
+
+const moduleName = 'managerHubApp';
+
+angular
+  .module(
+    moduleName,
+    [
+      atInternet,
+      dashboard,
+      errorPage,
+      ngOvhUiRouterLineProgress,
+      ngUiRouterBreadcrumb,
+      'oui',
+      ovhManagerCore,
+      ovhManagerHub,
+      ovhManagerNavbar,
+      ovhManagerOrderTracking,
+      'pascalprecht.translate',
+      uiRouter,
+      ...get(__NG_APP_INJECTIONS__, Environment.getRegion(), []),
+    ].filter(isString),
+  )
+  .controller('HubController', controller)
+  .config(
+    /* @ngInject */ ($locationProvider) => $locationProvider.hashPrefix(''),
+  )
+  .config(routing)
+  .run(
+    /* @ngInject */ ($anchorScroll, $rootScope, $translate, $transitions) => {
+      $transitions.onBefore({ to: 'app.**' }, () => $translate.refresh());
+      $transitions.onSuccess({}, () => $anchorScroll('hub-scroll-top'));
+
+      $transitions.onSuccess({ to: 'error' }, () => {
+        $rootScope.$emit('ovh::sidebar::hide');
+      });
+    },
+  )
+  .run(($translate) => {
+    moment.locale(head($translate.use().split('_')));
+  })
+  .run(
+    /* @ngInject */ ($state) => {
+      $state.defaultErrorHandler((error) => {
+        if (error.type === RejectType.ERROR) {
+          $state.go(
+            'error',
+            {
+              detail: {
+                message: get(error.detail, 'data.message'),
+                code: has(error.detail, 'headers')
+                  ? error.detail.headers('x-ovh-queryId')
+                  : null,
+              },
+            },
+            { location: false },
+          );
+        }
+      });
+    },
+  )
+  .run(
+    /* @ngInject */ ($rootScope, $transitions) => {
+      $transitions.onSuccess({ to: 'error' }, () => {
+        $rootScope.$emit('ovh::sidebar::hide');
+      });
+    },
+  )
+  .run(
+    /* @ngInject */ (ssoAuthentication) => {
+      if (!BILLING_REDIRECTIONS.includes(window.location.href)) {
+        ssoAuthentication.login();
+      }
+    },
+  )
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($rootScope, $transitions) => {
+      const unregisterHook = $transitions.onSuccess({}, () => {
+        detachPreloader();
+        unregisterHook();
+      });
+    },
+  );
+
+export default moduleName;
