--- packages/manager/apps/hub/src/components/manager-preload/manager-preload.less
@@ -1,23 +1,0 @@
-@import 'ovh-ui-kit/packages/oui-color/_variables.less';
-
-#managerPreload {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #fff;
-  z-index: 9999;
-
-  hr {
-    border-top: 2px solid @oui-color-primary;
-    width: 30%;
-    margin: auto;
-  }
-
-  &.manager-preload-hide {
-    transition: opacity 800ms 1s, visibility 1800ms;
-    opacity: 0;
-    visibility: hidden !important;
-  }
-}
