--- packages/manager/apps/hub/src/components/manager-preload/translations/Messages_es_ES.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Bienvenido/a al",
-  "preload_manager_ovh": "Área de cliente de OVHcloud"
-}
