--- packages/manager/apps/hub/src/components/manager-preload/translations/Messages_fi_FI.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Welcome to the",
-  "preload_manager_ovh": "OVHcloud Control Panel"
-}
