--- packages/manager/apps/hub/src/components/manager-preload/translations/Messages_fr_CA.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Bienvenue sur votre",
-  "preload_manager_ovh": "Espace client OVHcloud"
-}
