+++ packages/manager/apps/hub/src/index.html
@@ -11,34 +11,10 @@
         <title>OVHcloud</title>
     </head>
     <body
-        data-ng-app="managerHubApp"
-        data-ng-strict-di
         data-ng-controller="HubController as $ctrl"
         class="h-100 overflow-hidden"
         id="hub-scroll-top"
     >
-        <div
-            id="managerPreload"
-            class="d-flex text-center"
-            data-ng-class="managerPreloadHide"
-        >
-            <div class="m-auto">
-                <p
-                    class="lead mb-3"
-                    data-ng-if="('preload_welcome' | translate) !== 'preload_welcome'"
-                    data-translate="preload_welcome"
-                ></p>
-                <hr class="my-0" />
-                <h1
-                    class="mt-3 mb-5"
-                    data-ng-if="('preload_manager_ovh' | translate) !== 'preload_manager_ovh'"
-                    data-translate="preload_manager_ovh"
-                ></h1>
-                <div class="mt-5">
-                    <oui-spinner data-size="l"></oui-spinner>
-                </div>
-            </div>
-        </div>
         <ovh-manager-navbar
             data-navbar-options="{
                 universe: 'hub',
