+++ packages/manager/apps/hub/webpack.config.js
@@ -1,4 +1,3 @@
-const _ = require('lodash');
 const fs = require('fs');
 const glob = require('glob');
 const path = require('path');
@@ -17,20 +16,21 @@ function readNgAppInjections(file) {
   return injections;
 }
 
-function getNgAppInjections(region) {
-  const injections = [
-    ...readNgAppInjections(`./.extras-${region}/ng-app-injections`),
-    ...readNgAppInjections('./.extras/ng-app-injections'),
-  ];
+function getNgAppInjections(regions) {
+  return regions.reduce((ngAppInjections, region) => {
+    const injections = [
+      ...readNgAppInjections(`./.extras-${region}/ng-app-injections`),
+      ...readNgAppInjections('./.extras/ng-app-injections'),
+    ];
 
-  const ngAppInjections = injections.map((val) => `'${val}'`).join(',');
-
-  return ngAppInjections || 'null';
+    return {
+      ...ngAppInjections,
+      [region]: JSON.stringify(injections),
+    };
+  }, {});
 }
 
 module.exports = (env = {}) => {
-  const REGION = _.upperCase(env.region || process.env.REGION || 'EU');
-
   const { config } = webpackConfig(
     {
       template: './src/index.html',
@@ -46,18 +46,16 @@ module.exports = (env = {}) => {
         ],
       },
     },
-    REGION ? Object.assign(env, { region: REGION }) : env,
+    env,
   );
 
   // Extra config files
   const extras = glob.sync(`./.extras/**/*.js`);
-  const extrasRegion = glob.sync(`./.extras-${REGION}/**/*.js`);
 
   return merge(config, {
     entry: {
       main: path.resolve('./src/index.js'),
       ...(extras.length > 0 ? { extras } : {}),
-      ...(extrasRegion.length > 0 ? { extrasRegion } : {}),
     },
     output: {
       path: path.join(__dirname, 'dist'),
@@ -76,8 +74,7 @@ module.exports = (env = {}) => {
         __NODE_ENV__: process.env.NODE_ENV
           ? `'${process.env.NODE_ENV}'`
           : '"development"',
-        __WEBPACK_REGION__: `'${REGION}'`,
-        __NG_APP_INJECTIONS__: getNgAppInjections(REGION),
+        __NG_APP_INJECTIONS__: getNgAppInjections(['EU', 'CA', 'US']),
       }),
     ],
   });
